RandomIdeasApp completed

+++ routes/ideas.js
@@ -75,17 +75,30 @@ router.put("/:id", async (req, res) => {
   // idea.tag = req.body.tag || idea.tag;
   // res.json({ success: true, data: idea });
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag,
+    const idea = await Idea.findById(req.params.id);
+
+    //match the username
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag,
+          },
         },
-      },
-      { new: true }
-    );
-    res.json({ success: true, data: updatedIdea });
+        { new: true }
+      );
+      return res.json({ success: true, data: updatedIdea });
+    }
+
+    //username do not match
+    res
+      .status(403)
+      .json({
+        success: false,
+        error: "You are not authorized to update this resource",
+      });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: "something went wrong" });
@@ -105,8 +118,19 @@ router.delete("/:id", async (req, res) => {
   // ideas.splice(index, 1);
   // res.json({ success: true, data: {} });
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+
+    //match the username
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    }
+
+    //username do not match
+    res.status(403).json({
+      success: false,
+      error: "you are not authorized to delete this resource",
+    });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: "something went wrong" });
